requirements.txt
django>=5.2.7
wagtail
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pymysql
psycopg2-binary
gunicorn
